- Add: extracting zip files and send containing pdf files to lexoffice (#3) - Add: ENV "TEST_MODE" for testing - Change Dockerfile for zip Support

+++ index.php
@@ -9,12 +9,13 @@ if (file_exists('.env')) {
 //Check .env Variables
 function setEnvDefault($envName, $envRequired, $envDefault = '')
 {
-    if ( (getenv($envName) == false && !isset($_ENV[$envName])) && $envRequired === true) {
+    if ((getenv($envName) == false && !isset($_ENV[$envName])) && $envRequired === true) {
         throw new Exception('ENV missing: ' . $envName);
     } else if (!isset($_ENV[$envName])) {
         $_ENV[$envName] = $envDefault;
     }
 }
+
 try {
     setEnvDefault('LO_API_KEY', true);
     setEnvDefault('IMAP_HOST', true);
@@ -26,8 +27,9 @@ try {
     setEnvDefault('DELETE_MAIL', false, 'false');
     setEnvDefault('MOVE_MAIL', false, 'false');
     setEnvDefault('MOVE_MAIL_PATH', true, 'INBOX.Lexoffice');
+    setEnvDefault('TEST_MODE', false, 'false');
 } catch (Exception $e) {
-    die($e->getMessage().PHP_EOL);
+    die($e->getMessage() . PHP_EOL);
 }
 
 $attachmetsDir = __DIR__ . DIRECTORY_SEPARATOR . 'attachments';
@@ -50,6 +52,10 @@ $mailbox->setConnectionArgs(
     CL_EXPUNGE // don't do non-secure authentication
 );
 
+if ($_ENV['TEST_MODE'] == 'true') {
+    echo '### TEST MODE ON ###' . PHP_EOL;
+}
+
 while (true) {
 
     try {
@@ -67,32 +73,65 @@ while (true) {
     foreach ($mailsIds as $mailId) {
         $mail = $mailbox->getMail($mailId);
 
-        echo '[MAIL] Found mail from '.$mail->fromAddress.PHP_EOL;
+        echo '[MAIL] Found mail from ' . $mail->fromAddress . PHP_EOL;
 
         if ($mail->hasAttachments()) {
 
             $attachments = $mail->getAttachments();
 
-            echo '[INFO] Found '. count($attachments). ' attachments'.PHP_EOL;
+            echo '[INFO] Found ' . count($attachments) . ' attachments' . PHP_EOL;
+
 
             //send to lexoffice :)
             foreach ($attachments as $attachment) {
-                rename($attachment->filePath, str_replace('.bin', '.pdf', $attachment->filePath));
-                $attachmentPath = str_replace('.bin', '.pdf', $attachment->filePath);
-                exec('curl https://api.lexoffice.io/v1/files -X POST -H "Authorization: Bearer '.$_ENV['LO_API_KEY'].'" -H "Content-Type: multipart/form-data" -H "Accept: application/json" -F "file=@' . $attachmentPath . '" -F "type=voucher"');
+
+                if ($attachment->subtype == 'ZIP') {
+                    echo '[INFO] Found zip-File...extract necessary' . PHP_EOL;
+                    $zip = new ZipArchive;
+                    $extractTo = str_replace('.bin', '', $attachment->filePath);
+                    if ($zip->open($attachment->filePath) === TRUE) {
+                        $zip->extractTo($extractTo . DIRECTORY_SEPARATOR);
+                        $zip->close();
+                        echo '[INFO] zip-File successfully extracted' . PHP_EOL;
+
+                        $files = scandir($extractTo);
+                        foreach ($files as $file) {
+                            if (str_contains($file, '.pdf')) {
+                                $attachmentPath = $extractTo . DIRECTORY_SEPARATOR . $file;
+                                echo '[INFO] send extracted file: ' . $attachmentPath . PHP_EOL;
+                                if ($_ENV['TEST_MODE'] == 'false') {
+                                    exec('curl https://api.lexoffice.io/v1/files -X POST -H "Authorization: Bearer ' . $_ENV['LO_API_KEY'] . '" -H "Content-Type: multipart/form-data" -H "Accept: application/json" -F "file=@' . $attachmentPath . '" -F "type=voucher"');
+                                }
+                            }
+                        }
+                        rename($attachment->filePath, str_replace('.bin', '.zip', $attachment->filePath));
+
+                    } else {
+                        echo '[ERROR] Extracting failed...' . PHP_EOL;
+                    }
+                } else {
+                    rename($attachment->filePath, str_replace('.bin', '.pdf', $attachment->filePath));
+                    $attachmentPath = str_replace('.bin', '.pdf', $attachment->filePath);
+
+                    if ($_ENV['TEST_MODE'] == 'false') {
+                        exec('curl https://api.lexoffice.io/v1/files -X POST -H "Authorization: Bearer ' . $_ENV['LO_API_KEY'] . '" -H "Content-Type: multipart/form-data" -H "Accept: application/json" -F "file=@' . $attachmentPath . '" -F "type=voucher"');
+                    }
+                }
             }
         } else {
             echo '[INFO] Mail has no attachments' . PHP_EOL;
         }
 
-        if ($_ENV['DELETE_MAIL'] == 'true') {
-            $mailbox->deleteMail($mail->id);
-            echo '[INFO] Mail deleted'.PHP_EOL;
-            $mailbox->expungeDeletedMails();
-        }
-        if ($_ENV['MOVE_MAIL'] == 'true') {
-            $mailbox->moveMail($mail->id, $_ENV['MOVE_MAIL_PATH']);
-            echo '[INFO] Mail moved to: '.$_ENV['MOVE_MAIL_PATH'].PHP_EOL;
+        if ($_ENV['TEST_MODE'] == 'false') {
+            if ($_ENV['DELETE_MAIL'] == 'true') {
+                $mailbox->deleteMail($mail->id);
+                echo '[INFO] Mail deleted' . PHP_EOL;
+                $mailbox->expungeDeletedMails();
+            }
+            if ($_ENV['MOVE_MAIL'] == 'true') {
+                $mailbox->moveMail($mail->id, $_ENV['MOVE_MAIL_PATH']);
+                echo '[INFO] Mail moved to: ' . $_ENV['MOVE_MAIL_PATH'] . PHP_EOL;
+            }
         }
     }
 }
